fix memory leak, pixel sharpness, xp cap and doubled map extension

+++ src/gameplay/player.h
@@ -18,7 +18,7 @@ typedef struct {
     uint8_t  equipped[EQUIP_SLOTS]; /* indexed by EquipSlot; 0xFF = empty */
     uint8_t  skills[SKILL_MAX];
     uint8_t  level;
-    uint8_t  xp;
+    uint16_t xp;
     uint8_t  skillPoints;
     uint8_t  hp;
     uint8_t  intelligence;
